+++ flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/consumer/InputChannel.java
@@ -319,7 +319,7 @@ public abstract class InputChannel {
         return false;
     }
 
-    public void notifyRequiredSegmentId(long segmentId) {}
+    public void notifyRequiredSegmentId(int segmentId) {}
 
     // ------------------------------------------------------------------------
